Move my profile edit page component (#178)
